+++ packages/powersync_flutter_libs/android/src/main/java/com/powersync/powersync_flutter_libs/PowersyncFlutterLibsPlugin.java
@@ -1,0 +1,13 @@
+package com.powersync.powersync_flutter_libs;
+
+import androidx.annotation.NonNull;
+
+import io.flutter.embedding.engine.plugins.FlutterPlugin;
+
+public class PowersyncFlutterLibsPlugin implements FlutterPlugin { 
+    @Override
+    public void onAttachedToEngine(@NonNull FlutterPluginBinding flutterPluginBinding) { }
+
+    @Override
+    public void onDetachedFromEngine(@NonNull FlutterPluginBinding binding) { }
+}