+++ packages/powersync_flutter_libs/ios/Classes/PowersyncFlutterLibsPlugin.h
@@ -1,0 +1,4 @@
+#import <Flutter/Flutter.h>
+
+@interface PowersyncFlutterLibsPlugin : NSObject <FlutterPlugin>
+@end