+++ packages/powersync_flutter_libs/lib/powersync_flutter_libs.dart
@@ -1,0 +1,1 @@
+library powersync_flutter_libs;